Refactor Bigtable schema registry classes
