+++ caraml-store-serving/src/main/java/dev/caraml/serving/store/bigtable/BigTableSchemaRegistry.java
@@ -6,57 +6,14 @@ import com.google.cloud.bigtable.data.v2.models.Row;
 import com.google.cloud.bigtable.data.v2.models.RowCell;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.google.common.collect.Iterables;
 import com.google.protobuf.ByteString;
-import java.util.concurrent.ExecutionException;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericDatumReader;
 import org.apache.avro.generic.GenericRecord;
 
-public class BigTableSchemaRegistry {
+public class BigTableSchemaRegistry extends BaseSchemaRegistry {
   private final BigtableDataClient client;
-  private final LoadingCache<SchemaReference, GenericDatumReader<GenericRecord>> cache;
-
-  private static String COLUMN_FAMILY = "metadata";
-  private static String QUALIFIER = "avro";
-  private static String KEY_PREFIX = "schema#";
-
-  public static class SchemaReference {
-    private final String tableName;
-    private final ByteString schemaHash;
-
-    public SchemaReference(String tableName, ByteString schemaHash) {
-      this.tableName = tableName;
-      this.schemaHash = schemaHash;
-    }
-
-    public String getTableName() {
-      return tableName;
-    }
-
-    public ByteString getSchemaHash() {
-      return schemaHash;
-    }
-
-    @Override
-    public int hashCode() {
-      int result = tableName.hashCode();
-      result = 31 * result + schemaHash.hashCode();
-      return result;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (o == null || getClass() != o.getClass()) return false;
-
-      SchemaReference that = (SchemaReference) o;
-
-      if (!tableName.equals(that.tableName)) return false;
-      return schemaHash.equals(that.schemaHash);
-    }
-  }
 
   public BigTableSchemaRegistry(BigtableDataClient client) {
     this.client = client;
@@ -67,17 +24,8 @@ public class BigTableSchemaRegistry {
     cache = CacheBuilder.newBuilder().build(schemaCacheLoader);
   }
 
-  public GenericDatumReader<GenericRecord> getReader(SchemaReference reference) {
-    GenericDatumReader<GenericRecord> reader;
-    try {
-      reader = this.cache.get(reference);
-    } catch (ExecutionException | CacheLoader.InvalidCacheLoadException e) {
-      throw new RuntimeException(String.format("Unable to find Schema"), e);
-    }
-    return reader;
-  }
-
-  private GenericDatumReader<GenericRecord> loadReader(SchemaReference reference) {
+  @Override
+  public GenericDatumReader<GenericRecord> loadReader(SchemaReference reference) {
     Row row =
         client.readRow(
             reference.getTableName(),
