+++ caraml-store-serving/src/main/java/dev/caraml/serving/store/bigtable/HBaseSchemaRegistry.java
@@ -2,11 +2,9 @@ package dev.caraml.serving.store.bigtable;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.google.protobuf.ByteString;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.concurrent.ExecutionException;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericDatumReader;
 import org.apache.avro.generic.GenericRecord;
@@ -17,49 +15,8 @@ import org.apache.hadoop.hbase.client.Get;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.Table;
 
-public class HBaseSchemaRegistry {
+public class HBaseSchemaRegistry extends BaseSchemaRegistry {
   private final Connection hbaseClient;
-  private final LoadingCache<SchemaReference, GenericDatumReader<GenericRecord>> cache;
-
-  private static String COLUMN_FAMILY = "metadata";
-  private static String QUALIFIER = "avro";
-  private static String KEY_PREFIX = "schema#";
-
-  public static class SchemaReference {
-    private final String tableName;
-    private final ByteString schemaHash;
-
-    public SchemaReference(String tableName, ByteString schemaHash) {
-      this.tableName = tableName;
-      this.schemaHash = schemaHash;
-    }
-
-    public String getTableName() {
-      return tableName;
-    }
-
-    public ByteString getSchemaHash() {
-      return schemaHash;
-    }
-
-    @Override
-    public int hashCode() {
-      int result = tableName.hashCode();
-      result = 31 * result + schemaHash.hashCode();
-      return result;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (o == null || getClass() != o.getClass()) return false;
-
-      SchemaReference that = (SchemaReference) o;
-
-      if (!tableName.equals(that.tableName)) return false;
-      return schemaHash.equals(that.schemaHash);
-    }
-  }
 
   public HBaseSchemaRegistry(Connection hbaseClient) {
     this.hbaseClient = hbaseClient;
@@ -70,17 +27,8 @@ public class HBaseSchemaRegistry {
     cache = CacheBuilder.newBuilder().build(schemaCacheLoader);
   }
 
-  public GenericDatumReader<GenericRecord> getReader(SchemaReference reference) {
-    GenericDatumReader<GenericRecord> reader;
-    try {
-      reader = this.cache.get(reference);
-    } catch (ExecutionException | CacheLoader.InvalidCacheLoadException e) {
-      throw new RuntimeException(String.format("Unable to find Schema"), e);
-    }
-    return reader;
-  }
-
-  private GenericDatumReader<GenericRecord> loadReader(SchemaReference reference) {
+  @Override
+  public GenericDatumReader<GenericRecord> loadReader(SchemaReference reference) {
     try {
       Table table = this.hbaseClient.getTable(TableName.valueOf(reference.getTableName()));
